+++ src/algebra/builtinAlgebras.ts
@@ -1,4 +1,6 @@
+import { PGA_2D_INTERPRETATION } from '../geometry/pga2Interpretation'
 import { VGA_2D_INTERPRETATION } from '../geometry/vga2Interpretation'
+import { PGA_2D_VISUALIZER } from '../visualization/pga2Visualizer'
 import { VGA_2D_VISUALIZER } from '../visualization/vga2Visualizer'
 import { createAlgebraRegistry, type AlgebraRegistry } from './registry'
 import { PGA_DEFINITION } from './pgaDefinition'
@@ -10,6 +12,8 @@ export function createBuiltinAlgebraRegistry(): AlgebraRegistry {
   registry.register(VGA_DEFINITION)
   registry.register(PGA_DEFINITION)
   registry.registerInterpretation(VGA_2D_INTERPRETATION as unknown as Parameters<AlgebraRegistry['registerInterpretation']>[0])
+  registry.registerInterpretation(PGA_2D_INTERPRETATION as unknown as Parameters<AlgebraRegistry['registerInterpretation']>[0])
   registry.registerVisualizer(VGA_2D_VISUALIZER)
+  registry.registerVisualizer(PGA_2D_VISUALIZER)
   return registry
 }